requirements.txt
Pillow==3.4.2
requests==2.12.3
wsgiref==0.1.2
